+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 1.0.2.0
+
+* Bundled record pattern field selectors with the X_<general>
+  type, for easier import.  Users may need to import both
+  X_name(..), and the T_specific type synonym.  Made sure
+  this is reflected clearly in the documentation.
+
+## 1.0.1.0
+
+* Exported SVCB modules that should have been public
+
 ## 1.0.0.0
 
 First public release on Hackage.  See [`README.md`](README.md) for